requirements.txt
# Core dependencies
pytest
python-dotenv
litellm

# Development dependencies
unittest-mock